separate Base and sysimg, and put less code in top-level files
